Fix crypto.subtle setup for latest jsdom implementation
A newer version of jsdom has a base setup for the crypto object, but it
doesn't include the crypto.subtle functions. We were replacing the top-
level crypto object, but the jsdom implementation behaves as the
browser, which is to say that if you try to replace the top-level crypto
object the attempt is silently ignored. As such, to provide access to
`crypto.subtle` in tests, we have to set the subproperty directly.